feat(Pytorch): Add an example of building an MLP from a pipeline
